move context and other updates
